fix(pr-132): resolve Copilot review comments
- Tests: retarget selfCorrection, yieldAnalysis, sessionQuality, and
  wasteOptimizer to import from canonical tests/usage/fixtures/turn.ts;
  delete duplicate tests/usage/_fixtures.ts
- SettingsPage: fix comingSoon flags — only "Data & Privacy" is truly
  unimplemented; all other sections already have real components; remove
  "A future wave will wire it." from the not-wired badge tooltip
- CHANGELOG: list all 6 feature flags whose descriptions were cleaned
  (usageAnalytics, agentSkillIndexer, devServerControl, liveActivity,
  taskDispatcher, mcpSecurityScan)

212 test files, 2333 passing.
